fix: show SSO authentication error once
